+++ SasquatchMac/SasquatchMac/Constants.h
@@ -8,6 +8,7 @@ static NSString *const kMSUserIdKey = @"userId";
 static NSString *const kMSLogUrl = @"logUrl";
 static NSString *const kMSAppSecret = @"appSecret";
 static NSString *const kMSLogTag = @"[SasquatchMac]";
+static NSString *const kMSManualSessionTracker = @"kMSManualSessionTracker";
 static NSString *const kMSUserConfirmationKey = @"MSAppCenterCrashesUserConfirmation";
 
 static NSString *const kSASCustomizedUpdateAlertKey = @"kSASCustomizedUpdateAlertKey";